requirements.txt
dns-lexicon==3.8.8
dnspython==2.1.0
python-telegram-bot==13.7
requests==2.26.0